requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
fonttools==4.34.4
idna==3.3
imageio==2.19.2
joblib==1.1.0
kiwisolver==1.4.4
llvmlite==0.38.0
matplotlib==3.5.2
networkx==2.8.5
numba==0.55.1
numpy==1.21.6
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
opencv-python-headless==4.8.0.76
packaging==21.3
pandas==1.4.3
Pillow==9.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
requests==2.27.1
scikit-image==0.19.3
scipy==1.8.0
seaborn==0.11.2
six==1.16.0
sns==0.1
tifffile==2022.8.3
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113
tqdm==4.66.1
typing-extensions==4.2.0
urllib3==1.26.9